+++ note/models/user.py
@@ -10,6 +10,7 @@ class Users(UserMixin, db.Model):
     telephone = db.Column(db.String(11), nullable=False)
     username = db.Column(db.String(20), nullable=False)
     password = db.Column(db.String(100), nullable=False)
+    icon = db.Column(db.String(20), default='user.png')
 
     def __init__(self, *args, **kwargs):
         telephone = kwargs.get('telephone')
@@ -18,6 +19,7 @@ class Users(UserMixin, db.Model):
         self.telephone = telephone
         self.username = username
         self.password = generate_password_hash(password)
+        self.icon = 'user.png'
 
     def check_password(self, password):
         return check_password_hash(self.password, password)